Moves VaultConfig to client package
